svm: conformance: extract callbacks into shared module
