+++ svm/src/conformance/mod.rs
@@ -2,6 +2,7 @@
 
 #[cfg(feature = "conformance")]
 pub mod account_state;
+pub mod callback;
 #[cfg(feature = "conformance")]
 pub mod elf_loader;
 #[cfg(feature = "conformance")]
